# services/blackboard/openapi.json
{
  "components": {
    "schemas": {
...
      "CtfChallengeClaimRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "expected_revision"
        ],
        "title": "CtfChallengeClaimRequest",
        "type": "object"
      },
      "CtfChallengeCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "connection": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "requirements": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requirements"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "title"
        ],
        "title": "CtfChallengeCreateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Blackboard API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/tasks/{task_id}/ctf/artifacts/{artifact_id}": {
      "get": {
        "operationId": "artifact_api_tasks__task_id__ctf_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Artifact Api Tasks  Task Id  Ctf Artifacts  Artifact Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Artifact",
        "tags": [
          "ctf"
        ]
      }
    },
    "/api/tasks/{task_id}/ctf/challenges": {
      "get": {
        "operationId": "challenges_api_tasks__task_id__ctf_challenges_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Challenges Api Tasks  Task Id  Ctf Challenges Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Challenges",
        "tags": [
          "ctf"
        ]
      },
      "post": {
        "operationId": "create_challenge_api_tasks__task_id__ctf_challenges_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CtfChallengeCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Create Challenge Api Tasks  Task Id  Ctf Challenges Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Challenge",
        "tags": [
          "ctf"
        ]
      }
    },
    "/api/tasks/{task_id}/ctf/challenges/{challenge_id}": {
      "get": {
        "operationId": "challenge_api_tasks__task_id__ctf_challenges__challenge_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "challenge_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Challenge Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Challenge Api Tasks  Task Id  Ctf Challenges  Challenge Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Challenge",
        "tags": [
          "ctf"
        ]
      },
      "patch": {
        "operationId": "update_challenge_api_tasks__task_id__ctf_challenges__challenge_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "challenge_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Challenge Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CtfChallengeUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Update Challenge Api Tasks  Task Id  Ctf Challenges  Challenge Id  Patch"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Challenge",
        "tags": [
          "ctf"
        ]
      }
    },
    "/api/tasks/{task_id}/ctf/challenges/{challenge_id}/claim": {
      "post": {
        "operationId": "claim_challenge_api_tasks__task_id__ctf_challenges__challenge_id__claim_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "challenge_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Challenge Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CtfChallengeClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Claim Challenge Api Tasks  Task Id  Ctf Challenges  Challenge Id  Claim Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Challenge",
        "tags": [
          "ctf"
        ]
      }
    },
    "/api/tasks/{task_id}/ctf/challenges/{challenge_id}/help": {
      "post": {
        "operationId": "request_help_api_tasks__task_id__ctf_challenges__challenge_id__help_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "challenge_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Challenge Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CtfHelpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Request Help Api Tasks  Task Id  Ctf Challenges  Challenge Id  Help Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Help",
        "tags": [
          "ctf"
        ]
      }
    },
...
    "/api/tasks/{task_id}/ctf/challenges/{challenge_id}/verification/manual": {
      "post": {
        "operationId": "manual_verification_api_tasks__task_id__ctf_challenges__challenge_id__verification_manual_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "challenge_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Challenge Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CtfManualVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Manual Verification Api Tasks  Task Id  Ctf Challenges  Challenge Id  Verification Manual Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Manual Verification",
        "tags": [
          "ctf"
        ]
      }
    },
...
  }
}